Recompile mathlib.lib for 🪟 Use SSE2 instruction set  - SSE2 is the absolute minimum required for x86-64 and is enabled by default on 64-bit, so no need to manually specify to use the architecture (https://learn.microsoft.com/en-us/cpp/c-runtime-library/reference/set-sse2-enable?view=msvc-140)

+++ public/tier0/threadtools.h
@@ -1178,7 +1178,7 @@ public:
 	CThreadSpinRWLock()
 	{
 		COMPILE_TIME_ASSERT( sizeof( LockInfo_t ) == sizeof( int64 ) );
-		Assert( (int)this % 8 == 0 );
+		Assert( (intp)this % 8 == 0 );
 		m_lockInfo.m_writerId = 0;
 		m_lockInfo.m_nReaders = 0;
 		m_lockInfo.m_i64 = 0;
